switched to double precision + some refactoring

+++ PlaneSampling.hpp
@@ -9,14 +9,14 @@ class PlaneSampling {
     Mesh::fArray faces;
     PlaneSampling(std::string path);
 	// Get vertex and face number
-	const inline unsigned int vertNum() const { return verts.size()/2; }
-	const inline unsigned int faceNum() const { return faces.size()/3; }
+	const inline uint vertNum() const { return verts.size()/2; }
+	const inline uint faceNum() const { return faces.size()/3; }
 
-	const inline GLfloat cAttrib(unsigned int vertexId,
-		unsigned int attribOffset) const {
+	const inline double cAttrib(uint vertexId,
+		uint attribOffset) const {
 		return verts[2 * vertexId + attribOffset];
 	}
-	const inline GLuint cFacei(unsigned int faceId, unsigned int n) const {
+	const inline uint cFacei(uint faceId, uint n) const {
 		return faces[3 * faceId + n];
 	}
 };
